Creation of Java Project & Problem 1 Attempt 1
I created the java project in eclipse and worked on the first attempt
of problem 1 from project euler.  There is a test in the test package
and the classes are in the src package.  The classes run the actual
equation through the main function and the tests hold the base test
that is explained in the problem.